+++ lib/event_store/notifications/publisher.ex
@@ -12,7 +12,7 @@ defmodule EventStore.Notifications.Publisher do
   alias EventStore.Notifications.Notification
 
   defmodule State do
-    defstruct [:conn, :event_store, :schema, :serializer, :subscribe_to]
+    defstruct [:conn, :event_store, :schema, :serializer, :metadata_serializer, :subscribe_to]
 
     def new(opts) do
       %State{
@@ -20,6 +20,7 @@ defmodule EventStore.Notifications.Publisher do
         event_store: Keyword.fetch!(opts, :event_store),
         schema: Keyword.fetch!(opts, :schema),
         serializer: Keyword.fetch!(opts, :serializer),
+        metadata_serializer: Keyword.fetch!(opts, :metadata_serializer),
         subscribe_to: Keyword.fetch!(opts, :subscribe_to)
       }
     end
@@ -62,14 +63,20 @@ defmodule EventStore.Notifications.Publisher do
       to_stream_version: to_stream_version
     } = notification
 
-    %State{conn: conn, schema: schema, serializer: serializer} = state
+    %State{
+      conn: conn,
+      schema: schema,
+      serializer: serializer,
+      metadata_serializer: metadata_serializer
+    } = state
 
     count = to_stream_version - from_stream_version + 1
 
     try do
       case Storage.read_stream_forward(conn, stream_id, from_stream_version, count, schema: schema) do
         {:ok, events} ->
-          deserialized_events = deserialize_recorded_events(events, serializer)
+          deserialized_events =
+            deserialize_recorded_events(events, serializer, metadata_serializer)
 
           {stream_uuid, deserialized_events}
 
@@ -87,8 +94,8 @@ defmodule EventStore.Notifications.Publisher do
     end
   end
 
-  defp deserialize_recorded_events(recorded_events, serializer) do
-    Enum.map(recorded_events, &RecordedEvent.deserialize(&1, serializer))
+  defp deserialize_recorded_events(recorded_events, serializer, metadata_serializer) do
+    Enum.map(recorded_events, &RecordedEvent.deserialize(&1, serializer, metadata_serializer))
   end
 
   defp broadcast(event_store, stream_uuid, events) do